pyproject.toml
[tool.poetry]
name = "number_predictor"
version = "0.1.0"
description = "Predict numbers based on images"
authors = ["Dean <aswegen.d@buas.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = ">=1.20,<=1.24"
pillow = "^9.5.0"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
tensorflow = ">=2.7,<2.10.9"
tensorflow-io-gcs-filesystem ="0.31.0"
azureml = "^0.2.7"
azure-core = "^1.26.4"
azure-identity = "^1.13.0"
azureml-core = "^1.51.0"
azure-ai-ml = "^1.7.2"
mlflow = "^2.4.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
Sphinx = "^7.0.1"
pytest = "^7.3.1"
ipykernel = "^6.23.1"
pypiwin32 = "^223"
docker = "^6.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
